Implement a simple pass manager for Rocket Chip
